+++ src/debugger/crash_screen.c
@@ -1,63 +1,50 @@
 #include <ultra64.h>
 #include <PR/os_internal_error.h>
-#include <PR/os_system.h>
 #include <stdarg.h>
 #include <string.h>
-#include "config/config_debug.h"
 #include "buffers/framebuffers.h"
 #include "types.h"
-#include "game/puppyprint.h"
+#include "puppyprint.h"
 #include "audio/external.h"
 #include "farcall.h"
-#include "game/game_init.h"
-#include "game/main.h"
-#include "game/debug.h"
-#include "game/rumble_init.h"
-#include "game/printf.h"
-
-#include "crash_screen.h"
-#include "map_parser.h"
-#include "disasm.h"
-#include "assert.h"
-#include "stacktrace.h"
+#include "game_init.h"
+#include "main.h"
+#include "debug.h"
+#include "rumble_init.h"
 
 #include "sm64.h"
 
-extern u16 sRenderedFramebuffer;
-extern struct SequenceQueueItem sBackgroundMusicQueue[6];
-extern void audio_signal_game_loop_tick(void);
-extern void stop_sounds_in_continuous_banks(void);
-extern void read_controller_inputs(s32 threadID);
-extern char *strstr(char *, char *);
+#include "printf.h"
 
-static char *crashPageNames[] = {
-    [PAGE_SIMPLE] = "(Overview)",
-    [PAGE_CONTEXT] = "(Context)",
+#define X_KERNING 6
+
+enum crashPages {
+    PAGE_CONTEXT,
 #ifdef PUPPYPRINT_DEBUG
-    [PAGE_LOG] = "(Log)",
+    PAGE_LOG,
 #endif
-    [PAGE_STACKTRACE] = "(Stack Trace)",
-    [PAGE_DISASM] = "(Disassembly)",
-    [PAGE_ASSERTS] = "(Assert)",
+    PAGE_STACKTRACE,
+    PAGE_DISASM,
+    PAGE_ASSERTS,
+    PAGE_COUNT
 };
 
-static u8 sCrashScreenCharToGlyph[128] = {
-    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-     0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 14, 15,
-    16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31,
-    32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47,
-    48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63,
-    64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79,
-    80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95,
+u8 gCrashScreenCharToGlyph[128] = {
+    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
+    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 41, -1, -1, -1, 43, -1, -1, 37, 38, -1, 42,
+    -1, 39, 44, -1,  0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 36, -1, -1, -1, -1, 40, -1, 10,
+    11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32,
+    33, 34, 35, -1, -1, -1, -1, -1, -1, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22,
+    23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, -1, -1, -1, -1, -1,
 };
 
-static u32 sCrashScreenFont[GLYPH_HEIGHT * FONT_ROWS * 2 + 1] = {
+// A height of seven pixels for each Character * nine rows of characters + one row unused.
+u32 gCrashScreenFont[7 * 9 + 1] = {
     #include "textures/crash_custom/crash_screen_font.ia1.inc.c"
 };
 
-static u8 crashPage = 0;
-static u8 updateBuffer = TRUE;
+u8 crashPage = 0;
+u8 updateBuffer = TRUE;
 
 static char crashScreenBuf[0x200];
 
@@ -87,9 +74,14 @@ char *gFpcsrDesc[6] = {
     "Inexact operation",
 };
 
-static u32 sProgramPosition = 0;
-static u16 gCrashScreenTextColor = 0xFFFF;
-static struct {
+
+
+extern u64 osClockRate;
+extern far char *parse_map(u32 pc);
+extern far void map_data_init(void);
+extern far char *find_function_in_stack(u32 *sp);
+
+struct {
     OSThread thread;
     u64 stack[THREAD2_STACK / sizeof(u64)];
     OSMesgQueue mesgQueue;
@@ -99,14 +91,6 @@ static struct {
     u16 height;
 } gCrashScreen;
 
-static void set_text_color(u32 r, u32 g, u32 b) {
-    gCrashScreenTextColor = GPACK_RGBA5551(r, g, b, 255);
-}
-
-static void reset_text_color(void) {
-    gCrashScreenTextColor = 0xFFFF;
-}
-
 void crash_screen_draw_rect(s32 x, s32 y, s32 w, s32 h) {
     u16 *ptr;
     s32 i, j;
@@ -114,7 +98,8 @@ void crash_screen_draw_rect(s32 x, s32 y, s32 w, s32 h) {
     ptr = gCrashScreen.framebuffer + gCrashScreen.width * y + x;
     for (i = 0; i < h; i++) {
         for (j = 0; j < w; j++) {
-            *ptr = 0x0001;
+            // 0xe738 = 0b1110011100111000
+            *ptr = ((*ptr & 0xe738) >> 2) | 1;
             ptr++;
         }
         ptr += gCrashScreen.width - w;
@@ -128,27 +113,18 @@ void crash_screen_draw_glyph(s32 x, s32 y, s32 glyph) {
     u32 rowMask;
     s32 i, j;
 
-    if (glyph > 0x7F) return;
-
-    data = &sCrashScreenFont[((glyph&0xF)*GLYPH_HEIGHT * 2) + (glyph >= 64)];
-
+    data = &gCrashScreenFont[glyph / 5 * 7];
     ptr = gCrashScreen.framebuffer + gCrashScreen.width * y + x;
 
-    u16 color = gCrashScreenTextColor;
-
-    for (i = 0; i < GLYPH_HEIGHT; i++) {
-        bit = 0x80000000U >> ((glyph >> 4) * GLYPH_WIDTH);
+    for (i = 0; i < 7; i++) {
+        bit = 0x80000000U >> ((glyph % 5) * 6);
         rowMask = *data++;
-        data ++;
 
-        for (j = 0; j < (GLYPH_WIDTH); j++) {
-            if (bit & rowMask) {
-                *ptr = color;
-            }
-            ptr++;
+        for (j = 0; j < 6; j++) {
+            *ptr++ = (bit & rowMask) ? 0xffff : 1;
             bit >>= 1;
         }
-        ptr += gCrashScreen.width - (GLYPH_WIDTH);
+        ptr += gCrashScreen.width - 6;
     }
 }
 
@@ -176,7 +152,7 @@ void crash_screen_print_with_newlines(s32 x, s32 y, const s32 xNewline, const ch
                 xOffset = xNewline;
             }
 
-            glyph = sCrashScreenCharToGlyph[*ptr & 0x7f];
+            glyph = gCrashScreenCharToGlyph[*ptr & 0x7f];
 
             if (*ptr == '\n') {
                 y += 10;
@@ -209,7 +185,7 @@ void crash_screen_print(s32 x, s32 y, const char *fmt, ...) {
         ptr = crashScreenBuf;
 
         while (*ptr && size-- > 0) {
-            glyph = sCrashScreenCharToGlyph[*ptr & 0x7f];
+            glyph = gCrashScreenCharToGlyph[*ptr & 0x7f];
 
             if (glyph != 0xff) {
                 crash_screen_draw_glyph(x, y, glyph);
@@ -244,215 +220,143 @@ void crash_screen_print_fpcsr(u32 fpcsr) {
     s32 i;
     u32 bit = BIT(17);
 
-    crash_screen_print(100, 220, "FPCSR:%08XH", fpcsr);
+    crash_screen_print(30, 155, "FPCSR:%08XH", fpcsr);
     for (i = 0; i < 6; i++) {
         if (fpcsr & bit) {
-            crash_screen_print(222, 220, "(%s)", gFpcsrDesc[i]);
+            crash_screen_print(132, 155, "(%s)", gFpcsrDesc[i]);
             return;
         }
         bit >>= 1;
     }
 }
 
-void draw_crash_overview(OSThread *thread, s32 cause) {
-    __OSThreadContext *tc = &thread->context;
-
-    crash_screen_draw_rect(0, 20, 320, 240);
-
-    crash_screen_print(LEFT_MARGIN, 20, "Thread %d (%s)", thread->id, gCauseDesc[cause]);
-
-#ifdef DEBUG_EXPORT_SYMBOLS
-    symtable_info_t info = get_symbol_info(tc->pc);
-
-    crash_screen_print(LEFT_MARGIN, 40, "Crash at: %s", info.func == NULL ? "Unknown" : info.func);
-    if (info.line != -1) {
-        crash_screen_print(LEFT_MARGIN, 60, "File: %s", info.file);
-#ifdef DEBUG_EXPORT_ALL_LINES
-        // This line only shows the correct value if every line is in the sym file
-        crash_screen_print(LEFT_MARGIN, 72, "Line: %d", info.line);
-#endif // DEBUG_EXPORT_ALL_LINES
-    }
-#endif // DEBUG_EXPORT_SYMBOLS
-
-    crash_screen_print(LEFT_MARGIN, 84, "Address: 0x%08X", tc->pc);
-}
-
 void draw_crash_context(OSThread *thread, s32 cause) {
     __OSThreadContext *tc = &thread->context;
-    crash_screen_draw_rect(0, 20, 320, 240);
-    crash_screen_print(LEFT_MARGIN, 20, "Thread:%d (%s)", thread->id, gCauseDesc[cause]);
-    crash_screen_print(LEFT_MARGIN, 30, "PC:%08XH   SR:%08XH   VA:%08XH", tc->pc, tc->sr, tc->badvaddr);
+    crash_screen_draw_rect(25, 20, 270, 210);
+    crash_screen_print(30, 20, "THREAD:%d  (%s)", thread->id, gCauseDesc[cause]);
+    crash_screen_print(30, 30, "PC:%08XH   SR:%08XH   VA:%08XH", tc->pc, tc->sr, tc->badvaddr);
     osWritebackDCacheAll();
-#ifdef DEBUG_EXPORT_SYMBOLS
-    char *fname = parse_map(tc->pc, TRUE);
-    crash_screen_print(LEFT_MARGIN, 40, "Crash at: %s", fname == NULL ? "Unknown" : fname);
-#endif // DEBUG_EXPORT_SYMBOLS
-    crash_screen_print(LEFT_MARGIN,  52, "AT:%08XH   V0:%08XH   V1:%08XH", (u32) tc->at, (u32) tc->v0, (u32) tc->v1);
-    crash_screen_print(LEFT_MARGIN,  62, "A0:%08XH   A1:%08XH   A2:%08XH", (u32) tc->a0, (u32) tc->a1, (u32) tc->a2);
-    crash_screen_print(LEFT_MARGIN,  72, "A3:%08XH   T0:%08XH   T1:%08XH", (u32) tc->a3, (u32) tc->t0, (u32) tc->t1);
-    crash_screen_print(LEFT_MARGIN,  82, "T2:%08XH   T3:%08XH   T4:%08XH", (u32) tc->t2, (u32) tc->t3, (u32) tc->t4);
-    crash_screen_print(LEFT_MARGIN,  92, "T5:%08XH   T6:%08XH   T7:%08XH", (u32) tc->t5, (u32) tc->t6, (u32) tc->t7);
-    crash_screen_print(LEFT_MARGIN, 102, "S0:%08XH   S1:%08XH   S2:%08XH", (u32) tc->s0, (u32) tc->s1, (u32) tc->s2);
-    crash_screen_print(LEFT_MARGIN, 112, "S3:%08XH   S4:%08XH   S5:%08XH", (u32) tc->s3, (u32) tc->s4, (u32) tc->s5);
-    crash_screen_print(LEFT_MARGIN, 122, "S6:%08XH   S7:%08XH   T8:%08XH", (u32) tc->s6, (u32) tc->s7, (u32) tc->t8);
-    crash_screen_print(LEFT_MARGIN, 132, "T9:%08XH   GP:%08XH   SP:%08XH", (u32) tc->t9, (u32) tc->gp, (u32) tc->sp);
-    crash_screen_print(LEFT_MARGIN, 142, "S8:%08XH   RA:%08XH",            (u32) tc->s8, (u32) tc->ra);
-#ifdef DEBUG_EXPORT_SYMBOLS
-    fname = parse_map(tc->ra, TRUE);
-    crash_screen_print(LEFT_MARGIN, 152, "RA at: %s", fname == NULL ? "Unknown" : fname);
-#endif // DEBUG_EXPORT_SYMBOLS
-
+    if ((u32)parse_map != MAP_PARSER_ADDRESS) {
+        char *fname = parse_map(tc->pc);
+        crash_screen_print(30, 40, "CRASH AT: %s", fname == NULL ? "UNKNOWN" : fname);
+    }
+    crash_screen_print(30,  50, "AT:%08XH   V0:%08XH   V1:%08XH", (u32) tc->at, (u32) tc->v0, (u32) tc->v1);
+    crash_screen_print(30,  60, "A0:%08XH   A1:%08XH   A2:%08XH", (u32) tc->a0, (u32) tc->a1, (u32) tc->a2);
+    crash_screen_print(30,  70, "A3:%08XH   T0:%08XH   T1:%08XH", (u32) tc->a3, (u32) tc->t0, (u32) tc->t1);
+    crash_screen_print(30,  80, "T2:%08XH   T3:%08XH   T4:%08XH", (u32) tc->t2, (u32) tc->t3, (u32) tc->t4);
+    crash_screen_print(30,  90, "T5:%08XH   T6:%08XH   T7:%08XH", (u32) tc->t5, (u32) tc->t6, (u32) tc->t7);
+    crash_screen_print(30, 100, "S0:%08XH   S1:%08XH   S2:%08XH", (u32) tc->s0, (u32) tc->s1, (u32) tc->s2);
+    crash_screen_print(30, 110, "S3:%08XH   S4:%08XH   S5:%08XH", (u32) tc->s3, (u32) tc->s4, (u32) tc->s5);
+    crash_screen_print(30, 120, "S6:%08XH   S7:%08XH   T8:%08XH", (u32) tc->s6, (u32) tc->s7, (u32) tc->t8);
+    crash_screen_print(30, 130, "T9:%08XH   GP:%08XH   SP:%08XH", (u32) tc->t9, (u32) tc->gp, (u32) tc->sp);
+    crash_screen_print(30, 140, "S8:%08XH   RA:%08XH",            (u32) tc->s8, (u32) tc->ra);
     crash_screen_print_fpcsr(tc->fpcsr);
 
     osWritebackDCacheAll();
-    crash_screen_print_float_reg( 10, 170,  0, &tc->fp0.f.f_even);
-    crash_screen_print_float_reg(100, 170,  2, &tc->fp2.f.f_even);
-    crash_screen_print_float_reg(190, 170,  4, &tc->fp4.f.f_even);
-    crash_screen_print_float_reg( 10, 180,  6, &tc->fp6.f.f_even);
-    crash_screen_print_float_reg(100, 180,  8, &tc->fp8.f.f_even);
-    crash_screen_print_float_reg(190, 180, 10, &tc->fp10.f.f_even);
-    crash_screen_print_float_reg( 10, 190, 12, &tc->fp12.f.f_even);
-    crash_screen_print_float_reg(100, 190, 14, &tc->fp14.f.f_even);
-    crash_screen_print_float_reg(190, 190, 16, &tc->fp16.f.f_even);
-    crash_screen_print_float_reg( 10, 200, 18, &tc->fp18.f.f_even);
-    crash_screen_print_float_reg(100, 200, 20, &tc->fp20.f.f_even);
-    crash_screen_print_float_reg(190, 200, 22, &tc->fp22.f.f_even);
-    crash_screen_print_float_reg( 10, 210, 24, &tc->fp24.f.f_even);
-    crash_screen_print_float_reg(100, 210, 26, &tc->fp26.f.f_even);
-    crash_screen_print_float_reg(190, 210, 28, &tc->fp28.f.f_even);
-    crash_screen_print_float_reg( 10, 220, 30, &tc->fp30.f.f_even);
+    crash_screen_print_float_reg( 30, 170,  0, &tc->fp0.f.f_even);
+    crash_screen_print_float_reg(120, 170,  2, &tc->fp2.f.f_even);
+    crash_screen_print_float_reg(210, 170,  4, &tc->fp4.f.f_even);
+    crash_screen_print_float_reg( 30, 180,  6, &tc->fp6.f.f_even);
+    crash_screen_print_float_reg(120, 180,  8, &tc->fp8.f.f_even);
+    crash_screen_print_float_reg(210, 180, 10, &tc->fp10.f.f_even);
+    crash_screen_print_float_reg( 30, 190, 12, &tc->fp12.f.f_even);
+    crash_screen_print_float_reg(120, 190, 14, &tc->fp14.f.f_even);
+    crash_screen_print_float_reg(210, 190, 16, &tc->fp16.f.f_even);
+    crash_screen_print_float_reg( 30, 200, 18, &tc->fp18.f.f_even);
+    crash_screen_print_float_reg(120, 200, 20, &tc->fp20.f.f_even);
+    crash_screen_print_float_reg(210, 200, 22, &tc->fp22.f.f_even);
+    crash_screen_print_float_reg( 30, 210, 24, &tc->fp24.f.f_even);
+    crash_screen_print_float_reg(120, 210, 26, &tc->fp26.f.f_even);
+    crash_screen_print_float_reg(210, 210, 28, &tc->fp28.f.f_even);
+    crash_screen_print_float_reg( 30, 220, 30, &tc->fp30.f.f_even);
 }
 
 
 #ifdef PUPPYPRINT_DEBUG
 void draw_crash_log(void) {
     s32 i;
-    crash_screen_draw_rect(0, 20, 320, 210);
+    crash_screen_draw_rect(25, 20, 270, 210);
     osWritebackDCacheAll();
 #define LINE_HEIGHT (25 + ((LOG_BUFFER_SIZE - 1) * 10))
     for (i = 0; i < LOG_BUFFER_SIZE; i++) {
-        crash_screen_print(LEFT_MARGIN, (LINE_HEIGHT - (i * 10)), consoleLogTable[i]);
+        crash_screen_print(30, (LINE_HEIGHT - (i * 10)), consoleLogTable[i]);
     }
 #undef LINE_HEIGHT
 }
 #endif
 
+
+// prints any function pointers it finds in the stack format:
+// SP address: function name
 void draw_stacktrace(OSThread *thread, UNUSED s32 cause) {
     __OSThreadContext *tc = &thread->context;
+    u32 temp_sp = (tc->sp + 0x14);
 
-    crash_screen_draw_rect(0, 20, 320, 240);
-    crash_screen_print(LEFT_MARGIN, 25, "Stack Trace from %08X:", (u32) tc->sp);
-
-#if defined(DEBUG_EXPORT_SYMBOLS) && defined(DEBUG_FULL_STACK_TRACE)
-    // Current Func (EPC)
-    crash_screen_print(LEFT_MARGIN, 35, "%08X (%s)", tc->pc, parse_map(tc->pc, TRUE));
-
-    // Previous Func (RA)
-    u32 ra = tc->ra;
-    symtable_info_t info = get_symbol_info(ra);
-
-    crash_screen_print(LEFT_MARGIN, 45, "%08X (%s:%d)", ra, info.func, info.line);
+    crash_screen_draw_rect(25, 20, 270, 210);
+    crash_screen_print(30, 25, "STACK TRACE FROM %08X:", temp_sp);
+    if ((u32) parse_map == MAP_PARSER_ADDRESS) {
+        crash_screen_print(30, 35, "CURRFUNC: NONE");
+    } else {
+        crash_screen_print(30, 35, "CURRFUNC: %s", parse_map(tc->pc));
+    }
 
     osWritebackDCacheAll();
 
-    static u32 generated = 0;
+    for (int i = 0; i < 18; i++) {
+        if ((u32) find_function_in_stack == MAP_PARSER_ADDRESS) {
+            crash_screen_print(30, (45 + (i * 10)), "STACK TRACE DISABLED");
+            break;
+        } else {
+            if ((u32) find_function_in_stack == MAP_PARSER_ADDRESS) {
+                return;
+            }
 
-    if (stackTraceGenerated == FALSE) {
-        generated = generate_stack(thread);
-        stackTraceGenerated = TRUE;
+            char *fname = find_function_in_stack(&temp_sp);
+            if ((fname == NULL) || ((*(u32*)temp_sp & 0x80000000) == 0)) {
+                crash_screen_print(30, (45 + (i * 10)), "%08X: UNKNOWN", temp_sp);
+            } else {
+                crash_screen_print(30, (45 + (i * 10)), "%08X: %s", temp_sp, fname);
+            }
+        }
     }
-    for (u32 i = 0; i < generated; i++) {
-        crash_screen_print(LEFT_MARGIN, 55 + (i * 10), get_stack_entry(i));
-    }
-#else // defined(DEBUG_EXPORT_SYMBOLS) && defined(DEBUG_FULL_STACK_TRACE)
-    // simple stack trace
-    u32 sp = tc->sp;
-
-    for (int i = 0; i < STACK_LINE_COUNT; i++) {
-        crash_screen_print(LEFT_MARGIN, 55 + (i * 10), "%3d: %08X", i, *((u32*)(sp + (i * 4))));
-        crash_screen_print(120, 55 + (i * 10), "%3d: %08X", i + STACK_LINE_COUNT, *((u32*)(sp + ((i + STACK_LINE_COUNT) * 4))));
-    }
-#endif // defined(DEBUG_EXPORT_SYMBOLS) && defined(DEBUG_FULL_STACK_TRACE)
 }
 
+extern char *insn_disasm(u32 insn, u32 isPC);
+static u32 sProgramPosition = 0;
 void draw_disasm(OSThread *thread) {
     __OSThreadContext *tc = &thread->context;
+    // u32 insn = *(u32*)tc->pc;
 
-    crash_screen_draw_rect(0, 20, 320, 240);
+    crash_screen_draw_rect(25, 20, 270, 210);
     if (sProgramPosition == 0) {
         sProgramPosition = (tc->pc - 36);
     }
-    crash_screen_print(LEFT_MARGIN, 25, "Program Counter: %08X", sProgramPosition);
+    crash_screen_print(30, 25, "DISASM %08X", sProgramPosition);
     osWritebackDCacheAll();
 
-    int skiplines = 0;
-#ifdef DEBUG_EXPORT_SYMBOLS
-    int currline = 0;
-#endif // DEBUG_EXPORT_SYMBOLS
 
     for (int i = 0; i < 19; i++) {
         u32 addr = (sProgramPosition + (i * 4));
+        u32 toDisasm = *(u32*)(addr);
 
-        char *disasm = insn_disasm((InsnData *)addr);
-
-
-        if (disasm[0] == 0) {
-            crash_screen_print(LEFT_MARGIN + 22, 35 + (skiplines * 10) + (i * 10), "%08X", addr);
-        } else {
-#ifdef DEBUG_EXPORT_SYMBOLS
-            symtable_info_t info = get_symbol_info(addr);
-
-            if (info.func_offset == 0 && info.distance == 0 && currline != info.line) {
-                currline = info.line;
-                set_text_color(239, 196, 15);
-                crash_screen_print(LEFT_MARGIN, 35 + (skiplines * 10) + (i * 10), "<%s:>", info.func);
-                reset_text_color();
-                skiplines++;
-            }
-#ifndef DEBUG_EXPORT_ALL_LINES
-            // catch `jal` and `jalr` callsites
-            if (disasm[0] == 'j' && disasm[1] == 'a') {
-#endif // DEBUG_EXPORT_ALL_LINES
-                if (info.line != -1) {
-                    set_text_color(200, 200, 200);
-                    crash_screen_print(LEFT_MARGIN, 35 + (skiplines * 10) + (i * 10), "%d:", info.line);
-                    reset_text_color();
-                }
-#ifndef DEBUG_EXPORT_ALL_LINES
-            }
-#endif // DEBUG_EXPORT_ALL_LINES
-
-#endif // DEBUG_EXPORT_SYMBOLS
-            if (addr == tc->pc) {
-                set_text_color(255, 0, 0);
-            } else {
-                reset_text_color();
-            }
-            crash_screen_print(LEFT_MARGIN + 22, 35 + (skiplines * 10) + (i * 10), "%s", disasm);
-        }
-
+        crash_screen_print(30, (35 + (i * 10)), "%s", insn_disasm(toDisasm, (addr == tc->pc)));
     }
 
-    reset_text_color();
     osWritebackDCacheAll();
 }
 
 void draw_assert(UNUSED OSThread *thread) {
-    crash_screen_draw_rect(0, 20, 320, 240);
+    crash_screen_draw_rect(25, 20, 270, 210);
 
-    crash_screen_print(LEFT_MARGIN, 25, "Assert");
+    crash_screen_print(30, 25, "ASSERT PAGE");
 
     if (__n64Assert_Filename != NULL) {
-        crash_screen_print(LEFT_MARGIN, 35, "File: %s", __n64Assert_Filename);
-        crash_screen_print(LEFT_MARGIN, 45, "Line %d", __n64Assert_LineNum);
-        crash_screen_print(LEFT_MARGIN, 55, "Condition:");
-        crash_screen_print(LEFT_MARGIN, 65, "(%s)", __n64Assert_Condition);
-        if (__n64Assert_MessageBuf[0] != 0) {
-            crash_screen_print(LEFT_MARGIN, 75, "Message:");
-            crash_screen_print(LEFT_MARGIN, 85, " %s", __n64Assert_MessageBuf);
-        }
+        crash_screen_print(30, 45, "FILE: %s", __n64Assert_Filename);
+        crash_screen_print(30, 55, "LINE: %d", __n64Assert_LineNum);
+        crash_screen_print(30, 75, "MESSAGE:", __n64Assert_Message);
+        crash_screen_print_with_newlines(36, 85, 30, __n64Assert_Message);
     } else {
-        crash_screen_print(LEFT_MARGIN, 35, "No failed assert to report.");
+        crash_screen_print(30, 45, "No failed assert to report.");
     }
 
     osWritebackDCacheAll();
@@ -471,24 +375,19 @@ void draw_crash_screen(OSThread *thread) {
 
     if (gPlayer1Controller->buttonPressed & R_TRIG) {
         crashPage++;
-        if (crashPage == PAGE_ASSERTS && tc->cause != EXC_SYSCALL) crashPage++;
         updateBuffer = TRUE;
     }
     if (gPlayer1Controller->buttonPressed & (L_TRIG | Z_TRIG)) {
         crashPage--;
-        if (crashPage == PAGE_ASSERTS && tc->cause != EXC_SYSCALL) crashPage--;
         updateBuffer = TRUE;
     }
-
-    if (crashPage == PAGE_DISASM) {
-        if (gPlayer1Controller->buttonDown & D_CBUTTONS) {
-            sProgramPosition += 4;
-            updateBuffer = TRUE;
-        }
-        if (gPlayer1Controller->buttonDown & U_CBUTTONS) {
-            sProgramPosition -= 4;
-            updateBuffer = TRUE;
-        }
+    if (gPlayer1Controller->buttonDown & D_CBUTTONS) {
+        sProgramPosition += 4;
+        updateBuffer = TRUE;
+    }
+    if (gPlayer1Controller->buttonDown & U_CBUTTONS) {
+        sProgramPosition -= 4;
+        updateBuffer = TRUE;
     }
 
     if ((crashPage >= PAGE_COUNT) && (crashPage != 255)) {
@@ -496,13 +395,11 @@ void draw_crash_screen(OSThread *thread) {
     }
     if (crashPage == 255) {
         crashPage = (PAGE_COUNT - 1);
-        if (crashPage == PAGE_ASSERTS && tc->cause != EXC_SYSCALL) crashPage--;
     }
     if (updateBuffer) {
-        crash_screen_draw_rect(0, 0, 320, 20);
-        crash_screen_print(LEFT_MARGIN, 5, "Page:%02d %-22s L/Z: Left   R: Right", crashPage, crashPageNames[crashPage]);
+        crash_screen_draw_rect(25, 8, 270, 12);
+        crash_screen_print(30, 10, "Page:%02d                L/Z: Left   R: Right", crashPage);
         switch (crashPage) {
-            case PAGE_SIMPLE:     draw_crash_overview(thread, cause); break;
             case PAGE_CONTEXT:    draw_crash_context(thread, cause); break;
 #ifdef PUPPYPRINT_DEBUG
             case PAGE_LOG: 		  draw_crash_log(); break;
@@ -532,6 +429,12 @@ OSThread *get_crashed_thread(void) {
     return NULL;
 }
 
+extern u16 sRenderedFramebuffer;
+extern void audio_signal_game_loop_tick(void);
+extern void stop_sounds_in_continuous_banks(void);
+extern void read_controller_inputs(s32 threadID);
+extern struct SequenceQueueItem sBackgroundMusicQueue[6];
+
 void thread2_crash_screen(UNUSED void *arg) {
     OSMesg mesg;
     OSThread *thread = NULL;
@@ -544,6 +447,9 @@ void thread2_crash_screen(UNUSED void *arg) {
             thread = get_crashed_thread();
             gCrashScreen.framebuffer = (RGBA16 *) gFramebuffers[sRenderedFramebuffer];
             if (thread) {
+                if ((u32) map_data_init != MAP_PARSER_ADDRESS) {
+                    map_data_init();
+                }
                 gCrashScreen.thread.priority = 15;
                 stop_sounds_in_continuous_banks();
                 stop_background_music(sBackgroundMusicQueue[0].seqId);
@@ -552,10 +458,6 @@ void thread2_crash_screen(UNUSED void *arg) {
                 play_sound(SOUND_MARIO_WAAAOOOW, gGlobalSoundSource);
                 audio_signal_game_loop_tick();
                 crash_screen_sleep(200);
-                // If an assert happened, go straight to that page
-                if (thread->context.cause == EXC_SYSCALL) {
-                    crashPage = PAGE_ASSERTS;
-                }
                 continue;
             }
         } else {
